Move policy callback to serialize
